File Sorting Option for Image Command
The first thought that the map files with the highest number are the
newest and, therefore, should be drawn last turned out to be wrong.
Instead, it makes more sense to draw the maps in order from oldest to
newest using the timestamps.

Sorting map files by timestamps are now the default behavior, but this
can be changed back to number order with the new image command option.