Restoring Map Listing
This commit restores a slightly improved version of the map listing. It
is now possible to find map files recursively from subdirectories.

The library was improved to return an iterator object for read_maps
instead of a vector of loaded content.

Dimensions in map files are always overworld, which could be later fixed
by checking the map file path.