Creating Test Map Item
This commit adds a new `test` subcommand that is available with the
`test_tool` feature. This subcommand is for creating files to help test
the program. Currently, there is an option for creating test map items.

The test map item was added to the tests folder with the corresponding
image captured from the Minecraft game.

The library was improved to allow the serialization of map items, and
the map item reader was also improved to use the Read interface for
tighter code.