Making Map Images and Testing
An improved version of make_image was restored for the library with a
unit test to check that it is working correctly. Luckily palette has
stayed the same since Minecraft version 21w10a, which base colors we got
from the Minecraft Wiki.

The unit test for the make_image compares generated map image against
the reference image captured from the game using the apitrace. All
incorrect colors are collected and printed in the panic message with the
values they should be.

The palette module was improved to have a constant base palette from the
Minecraft Wiki, which can be used with the new version of the
generate_palette function.

A new error kind was added for the error module to report problems with
map items.