Making Map Image
This commit adds a new image tool command to create an image file from a
single map file. Output file optional as the following update is planned
to include a possibility to output image directly to the terminal with
the help of the viuer crate.

Commit also includes an improved way of formatting output messages,
recommended by the clippy tool.