Info Improvements and Dev Tools
A new pretty_dimension() method was added to the MapItem. This one tries
to detect dimension from the file path and uses the MapData
pretty_diemnsion() as a fallback.

Dev Tools are now enabled with the dev_tools feature. A new sub-command
was introduced to automatically create a version source file from the
Minecraft Wiki.

The Test Map tool was updated to allow setting a specific version for
the map file. The update_versions flag was removed because of the
above-mentioned new sub-command.

Info and List tool now use the new pretty_dimension() method from the
MapItem.

All subcommands were modified to use the same convention of run()
function that is called from the main function.

An optional output directory was added for the Images tool, but
implementation has yet to start.

The test map item documentation was updated with a short description of
how the test image is updated.