World Dimension from Map Items
World dimension is now taken from the map items instead of a directory
name. The directory name can still be used with the proper option. This
change is due to the realization that the test files we used got mixed
up when moving between versions and servers. Map items appear to have
correct dimension information after all.